feat: implement pqrs backend core cu-01 to cu-07 plus rf-12

+++ Proyecto-Final/backend/pqrs/src/main/java/co/edu/konrad/pqrs/infrastructure/persistencia/UsuarioJpaRepositorio.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public  interface UsuarioJpaRepositorio extends JpaRepository<UsuarioEntidad, Long> {
+public interface UsuarioJpaRepositorio extends JpaRepository<UsuarioEntidad, Integer> {
 
     Optional<UsuarioEntidad> findByEmail(String email);
